refactor: remove invokeApi usage
