--- src/lib/invokeApi.js
@@ -1,54 +1,0 @@
-import { ApiError, AuthenticationService } from '$lib/api';
-import { auth } from './auth.svelte.js';
-import { goto } from '$app/navigation';
-import { resolve } from '$app/paths';
-
-let refreshPromise = null;
-
-/**
- * Wraps an API service call with automatic 401 refresh-and-retry.
- * Usage: await invoke(() => VacancyService.getVacancies(...))
- */
-export async function invokeApi(operation) {
-	try {
-		return await operation();
-	} catch (error) {
-		if (!(error instanceof ApiError) || error.status !== 401) {
-			throw error;
-		}
-		await refreshToken();
-		return await operation();
-	}
-}
-
-/**
- * Performs a coalesced token refresh. Concurrent callers share a single
- * in-flight refresh promise. Exported so the svelte-query layer
- * (`authedQueryFn` in `queryClient.js`) reuses the same coalescing as
- * `invokeApi`, avoiding duplicate refresh calls while both coexist.
- */
-export async function refreshToken() {
-	if (refreshPromise) {
-		await refreshPromise;
-		return;
-	}
-	refreshPromise = doRefresh();
-	try {
-		await refreshPromise;
-	} finally {
-		refreshPromise = null;
-	}
-}
-
-async function doRefresh() {
-	try {
-		const response = await AuthenticationService.refresh();
-		auth.accessToken = response.access_token;
-	} catch (error) {
-		auth.clear();
-		// Capture current URL to return after re-authentication
-		const returnUrl = window.location.pathname + window.location.search;
-		await goto(resolve(`/login?returnUrl=${encodeURIComponent(returnUrl)}`));
-		throw error;
-	}
-}
